res changed the type of csArray<> indices and sizes from int to   size_t. CVS has been tagged with PRE_CSARRAY_SIZE_T_INDICES before   the change.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18434 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/demo/demo.h
@@ -68,7 +68,7 @@ public:
   bool message_error;
 
   int do_demo;
-  int selected_demo;
+  size_t selected_demo;
   csStringArray demos;
   int first_y;	// First y location where list of demo files start.
